requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
google-auth==2.48.0
google-genai==1.65.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.24.0
pyasn1==0.6.2
pyasn1_modules==4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.1.0
python-dotenv==1.2.1
python-multipart==0.0.6
regex==2026.2.19
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.354
langchain-community==0.0.21
langchain-groq==0.1.0
langchain-huggingface==0.0.1
langchain-qdrant==0.1.11
qdrant-client==2.7.2
sentence-transformers==2.2.2
pypdf==3.17.1
